+++ doc/releases.rst
@@ -5,6 +5,8 @@ Releases
     | Version: 1.21.0
     | Released: 2020-07-20
 
+    - Added *clone* argument to :class:`inform.InformantFactory`.
+
 **1.21 (2020-07-20)**:
     - Allow :class:`inform.ProgressBar` output to be suppressed.
     - Allow ``/`` to be overridden in :class:`inform.plural`